+++ swift_browser_ui_frontend/src/common/lang.js
@@ -66,6 +66,11 @@ let default_translations = {
       logOut: "Log Out",
       cscOrg: "CSC - IT Center For Science LTD",
       devel: "Developed by",
+      folderTabs: {
+        all: "All project's folders",
+        sharedFrom: "Shared from project",
+        sharedTo: "Shared to project",
+      },
       table: {
         name: "Name",
         objects: "Objects",
@@ -406,6 +411,11 @@ let default_translations = {
       logOut: "Kirjaudu ulos",
       cscOrg: "CSC – Tieteen Tietotekniikan Keskus Oy",
       devel: "kehittänyt",
+      folderTabs: {
+        all: "Kaikki projektin kansiot",
+        sharedFrom: "Projektista jaetut",
+        sharedTo: "Projektiin jaetut",
+      },
       table: {
         name: "Nimi",
         objects: "Objekteja",